microservices-demo/services/todo-service/requirements.txt
aio-pika==9.5.5
aiormq==6.8.1
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
exceptiongroup==1.3.0
fastapi==0.115.12
graphql-core==3.2.6
grpcio==1.73.0
grpcio-tools==1.73.0
grpclib==0.4.8
h11==0.16.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
motor==3.7.1
multidict==6.4.4
packaging==25.0
pamqp==3.3.0
propcache==0.3.2
protobuf==6.31.1
pydantic==2.11.6
pydantic_core==2.33.2
pymongo==4.13.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
strawberry-graphql==0.273.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
yarl==1.20.1
